+++ Assets/Scripts/BuffStatsFeatures/Core/Configs/Buffs/BerserkBuff.cs
@@ -3,10 +3,9 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "BerserkBuff", menuName = "Configs/Buffs/BerserkBuff")]
 public class BerserkBuff : BuffConfig
 {
-    [SerializeField] private float _damageMultiplier;
-    [SerializeField] private float _armorReduction;
-    [SerializeField] private float _duration;
-
+    [SerializeField] private int _damageMultiplier;
+    [SerializeField] private int _armorReduction;
+    [SerializeField] private int _duration;
 
 
     public override IBuff CreateBuff()
